Refactored and decentralized grid item components into screen-specific packages and simplified folder transition logic:
- Deleted the shared `GridItem.kt` component file, moving its logic into targeted implementations within the `folder` and `pager` screen packages.
- Introduced `InteractiveFolderApplicationInfoGridItem`, `InteractiveFolderShortcutInfoGridItem`, and `InteractiveFolderShortcutConfigGridItem` to handle specialized rendering within folder views.
- Implemented dynamic padding in folder items using `lerp` and a new `progress` parameter to support smoother opening/closing animations.
- Refactored `FolderScreen` and `PagerScreen` to remove unused parameters related to drag-and-drop state management, such as `isDragging` and `lockMovement`.
- Simplified the `InteractiveFolderGridItem` signature by removing folder-specific dragging callbacks (`onShowFolderWhenDragging`, `onUpdateFolderPopupBounds`).
- Updated `HomeUtil` to change the default folder preview grid dimensions from 3x3 to 2x2.
- Migrated shared UI logic for applications, widgets, and shortcuts into `pager/InteractiveGridItem.kt` to improve package encapsulation.
- Updated local IDE deployment target settings.
