--- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/component/GridItem.kt
@@ -1,1256 +1,0 @@
-/*
- *
- *   Copyright 2023 Einstein Blanco
- *
- *   Licensed under the GNU General Public License v3.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       https://www.gnu.org/licenses/gpl-3.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.eblan.launcher.feature.home.component
-
-import androidx.compose.animation.ExperimentalSharedTransitionApi
-import androidx.compose.animation.SharedTransitionScope
-import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.draw.drawWithContent
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.layer.drawLayer
-import androidx.compose.ui.graphics.rememberGraphicsLayer
-import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.layout.onGloballyPositioned
-import androidx.compose.ui.layout.positionInRoot
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.IntOffset
-import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.round
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.viewinterop.AndroidView
-import coil3.compose.AsyncImage
-import coil3.request.ImageRequest.Builder
-import coil3.request.addLastModifiedToFileCacheKey
-import coil3.size.Size
-import com.eblan.launcher.domain.model.GridItem
-import com.eblan.launcher.domain.model.GridItemData
-import com.eblan.launcher.domain.model.GridItemSettings
-import com.eblan.launcher.domain.model.MoveGridItemResult
-import com.eblan.launcher.feature.home.model.Drag
-import com.eblan.launcher.feature.home.model.GridItemSource
-import com.eblan.launcher.feature.home.model.SharedElementKey
-import com.eblan.launcher.feature.home.screen.pager.handleConflictingGridItem
-import com.eblan.launcher.feature.home.util.FOLDER_COLUMNS
-import com.eblan.launcher.feature.home.util.FOLDER_ROWS
-import com.eblan.launcher.feature.home.util.getHorizontalAlignment
-import com.eblan.launcher.feature.home.util.getVerticalArrangement
-import com.eblan.launcher.feature.home.util.handleDrag
-import com.eblan.launcher.feature.home.util.onDoubleTap
-import com.eblan.launcher.feature.home.util.onLongPress
-import com.eblan.launcher.ui.local.LocalAppWidgetHost
-import com.eblan.launcher.ui.local.LocalAppWidgetManager
-import com.eblan.launcher.ui.local.LocalLauncherApps
-import com.eblan.launcher.ui.local.LocalSettings
-import kotlinx.coroutines.launch
-
-@OptIn(ExperimentalSharedTransitionApi::class)
-@Composable
-internal fun SharedTransitionScope.InteractiveApplicationInfoGridItem(
-    modifier: Modifier = Modifier,
-    data: GridItemData.ApplicationInfo,
-    drag: Drag,
-    gridItem: GridItem,
-    gridItemSettings: GridItemSettings,
-    iconPackFilePaths: Map<String, String>,
-    isScrollInProgress: Boolean,
-    isSelected: Boolean,
-    isShowWhiteBox: Boolean,
-    isVisibleFolder: Boolean,
-    isVisibleOverlay: Boolean,
-    newGridItemSource: GridItemSource,
-    sharedElementKey: SharedElementKey,
-    statusBarNotifications: Map<String, Int>,
-    textColor: Color,
-    onDismissGridItemPopup: () -> Unit,
-    onOpenAppDrawer: () -> Unit,
-    onShowGridItemPopup: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onTapApplicationInfo: (
-        serialNumber: Long,
-        componentName: String,
-    ) -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
-) {
-    val launcherApps = LocalLauncherApps.current
-
-    val context = LocalContext.current
-
-    val settings = LocalSettings.current
-
-    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
-
-    var intSize by remember { mutableStateOf(IntSize.Zero) }
-
-    val graphicsLayer = rememberGraphicsLayer()
-
-    val scope = rememberCoroutineScope()
-
-    val horizontalAlignment =
-        getHorizontalAlignment(horizontalAlignment = gridItemSettings.horizontalAlignment)
-
-    val verticalArrangement =
-        getVerticalArrangement(verticalArrangement = gridItemSettings.verticalArrangement)
-
-    val maxLines = if (gridItemSettings.singleLineLabel) 1 else Int.MAX_VALUE
-
-    val icon = iconPackFilePaths[data.componentName] ?: data.icon
-
-    val hasNotifications =
-        statusBarNotifications[data.packageName] != null && (
-            statusBarNotifications[data.packageName]
-                ?: 0
-            ) > 0
-
-    val hasInteraction = isSelected && isVisibleOverlay
-
-    val isVisibleWhiteBox = isShowWhiteBox && isSelected && drag == Drag.Dragging
-
-    val alpha = if (hasInteraction) 0f else 1f
-
-    LaunchedEffect(key1 = drag) {
-        handleDrag(
-            drag = drag,
-            isSelected = isSelected,
-            isVisibleOverlay = isVisibleOverlay,
-            onUpdateIsDragging = onUpdateIsDragging,
-            onDismissGridItemPopup = onDismissGridItemPopup,
-        )
-    }
-
-    Column(
-        modifier = modifier
-            .pointerInput(key1 = drag) {
-                detectTapGestures(
-                    onDoubleTap = if (!isVisibleOverlay) {
-                        {
-                            onDoubleTap(
-                                context = context,
-                                doubleTap = gridItem.doubleTap,
-                                launcherApps = launcherApps,
-                                scope = scope,
-                                onOpenAppDrawer = onOpenAppDrawer,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onLongPress = if (!isVisibleOverlay) {
-                        {
-                            onLongPress(
-                                scope = scope,
-                                graphicsLayer = graphicsLayer,
-                                intOffset = intOffset,
-                                intSize = intSize,
-                                gridItemSource = newGridItemSource,
-                                sharedElementKey = sharedElementKey,
-                                gridItem = gridItem,
-                                onUpdateGridItemSource = onUpdateGridItemSource,
-                                onUpdateImageBitmap = onUpdateImageBitmap,
-                                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                onShowGridItemPopup = onShowGridItemPopup,
-                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onTap = if (!isVisibleOverlay) {
-                        {
-                            scope.launch {
-                                onTapApplicationInfo(
-                                    data.serialNumber,
-                                    data.componentName,
-                                )
-                            }
-                        }
-                    } else {
-                        null
-                    },
-                )
-            }
-            .swipeGestures(
-                swipeDown = gridItem.swipeDown,
-                swipeUp = gridItem.swipeUp,
-                onOpenAppDrawer = onOpenAppDrawer,
-            )
-            .fillMaxSize()
-            .padding(gridItemSettings.padding.dp)
-            .background(
-                color = Color(gridItemSettings.customBackgroundColor),
-                shape = RoundedCornerShape(size = gridItemSettings.cornerRadius.dp),
-            )
-            .whiteBox(
-                textColor = textColor,
-                visible = isVisibleWhiteBox && !isVisibleFolder,
-            ),
-        horizontalAlignment = horizontalAlignment,
-        verticalArrangement = verticalArrangement,
-    ) {
-        Box(
-            modifier = Modifier
-                .size(gridItemSettings.iconSize.dp)
-                .alpha(alpha),
-        ) {
-            AsyncImage(
-                model = Builder(LocalContext.current).data(data.customIcon ?: icon)
-                    .addLastModifiedToFileCacheKey(true)
-                    .size(Size.ORIGINAL)
-                    .build(),
-                contentDescription = null,
-                modifier = Modifier
-                    .matchParentSize()
-                    .drawWithContent {
-                        graphicsLayer.record {
-                            this@drawWithContent.drawContent()
-                        }
-
-                        drawLayer(graphicsLayer)
-                    }
-                    .onGloballyPositioned { layoutCoordinates ->
-                        intOffset = layoutCoordinates.positionInRoot().round()
-
-                        intSize = layoutCoordinates.size
-                    }
-                    .run {
-                        if (!hasInteraction) {
-                            sharedElementWithCallerManagedVisibility(
-                                rememberSharedContentState(
-                                    key = sharedElementKey,
-                                ),
-                                visible = !isScrollInProgress,
-                            )
-                        } else {
-                            this
-                        }
-                    },
-            )
-
-            if (settings.isNotificationAccessGranted() && hasNotifications) {
-                Box(
-                    modifier = Modifier
-                        .size((gridItemSettings.iconSize * 0.3).dp)
-                        .align(Alignment.TopEnd)
-                        .background(
-                            color = MaterialTheme.colorScheme.primary,
-                            shape = CircleShape,
-                        ),
-                )
-            }
-        }
-
-        if (gridItemSettings.showLabel) {
-            Text(
-                modifier = Modifier.alpha(alpha),
-                text = data.customLabel ?: data.label,
-                color = textColor,
-                textAlign = TextAlign.Center,
-                maxLines = maxLines,
-                fontSize = gridItemSettings.textSize.sp,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalSharedTransitionApi::class)
-@Composable
-internal fun SharedTransitionScope.InteractiveWidgetGridItem(
-    modifier: Modifier = Modifier,
-    data: GridItemData.Widget,
-    drag: Drag,
-    isScrollInProgress: Boolean,
-    isSelected: Boolean,
-    isShowWhiteBox: Boolean,
-    isVisibleOverlay: Boolean,
-    newGridItemSource: GridItemSource,
-    sharedElementKey: SharedElementKey,
-    textColor: Color,
-    gridItem: GridItem,
-    onDismissGridItemPopup: () -> Unit,
-    onShowGridItemPopup: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
-) {
-    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
-
-    var intSize by remember { mutableStateOf(IntSize.Zero) }
-
-    val appWidgetHost = LocalAppWidgetHost.current
-
-    val appWidgetManager = LocalAppWidgetManager.current
-
-    val appWidgetInfo = appWidgetManager.getAppWidgetInfo(appWidgetId = data.appWidgetId)
-
-    val graphicsLayer = rememberGraphicsLayer()
-
-    val scope = rememberCoroutineScope()
-
-    val hasInteraction = isSelected && isVisibleOverlay
-
-    val isVisibleWhiteBox = isShowWhiteBox && isSelected && drag == Drag.Dragging
-
-    val alpha = if (hasInteraction) 0f else 1f
-
-    LaunchedEffect(key1 = drag) {
-        handleDrag(
-            drag = drag,
-            isSelected = isSelected,
-            isVisibleOverlay = isVisibleOverlay,
-            onUpdateIsDragging = onUpdateIsDragging,
-            onDismissGridItemPopup = onDismissGridItemPopup,
-        )
-    }
-
-    Box(
-        modifier = modifier
-            .fillMaxSize()
-            .whiteBox(textColor = textColor, visible = isVisibleWhiteBox),
-    ) {
-        val commonModifier = Modifier
-            .matchParentSize()
-            .alpha(alpha)
-            .drawWithContent {
-                graphicsLayer.record {
-                    this@drawWithContent.drawContent()
-                }
-
-                drawLayer(graphicsLayer)
-            }
-            .onGloballyPositioned { layoutCoordinates ->
-                intOffset = layoutCoordinates.positionInRoot().round()
-
-                intSize = layoutCoordinates.size
-            }
-            .run {
-                if (!hasInteraction) {
-                    sharedElementWithCallerManagedVisibility(
-                        rememberSharedContentState(
-                            key = sharedElementKey,
-                        ),
-                        visible = !isScrollInProgress,
-                    )
-                } else {
-                    this
-                }
-            }
-
-        if (appWidgetInfo != null) {
-            AndroidView(
-                factory = {
-                    appWidgetHost.createView(
-                        appWidgetId = data.appWidgetId,
-                        appWidgetProviderInfo = appWidgetInfo,
-                    )
-                },
-                modifier = commonModifier,
-                update = { view ->
-                    if (!isVisibleOverlay) {
-                        view.setOnLongClickListener {
-                            onLongPress(
-                                scope = scope,
-                                graphicsLayer = graphicsLayer,
-                                intOffset = intOffset,
-                                intSize = intSize,
-                                gridItemSource = newGridItemSource,
-                                sharedElementKey = sharedElementKey,
-                                gridItem = gridItem,
-                                onUpdateGridItemSource = onUpdateGridItemSource,
-                                onUpdateImageBitmap = onUpdateImageBitmap,
-                                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                onShowGridItemPopup = onShowGridItemPopup,
-                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
-                            )
-
-                            true
-                        }
-                    }
-                },
-            )
-        } else {
-            AsyncImage(
-                model = data.preview ?: data.icon,
-                contentDescription = null,
-                modifier = commonModifier.pointerInput(key1 = drag) {
-                    detectTapGestures(
-                        onLongPress = if (!isVisibleOverlay) {
-                            {
-                                onLongPress(
-                                    scope = scope,
-                                    graphicsLayer = graphicsLayer,
-                                    intOffset = intOffset,
-                                    intSize = intSize,
-                                    gridItemSource = newGridItemSource,
-                                    sharedElementKey = sharedElementKey,
-                                    gridItem = gridItem,
-                                    onUpdateGridItemSource = onUpdateGridItemSource,
-                                    onUpdateImageBitmap = onUpdateImageBitmap,
-                                    onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                    onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                    onShowGridItemPopup = onShowGridItemPopup,
-                                    onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                                    onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
-                                )
-                            }
-                        } else {
-                            null
-                        },
-                    )
-                },
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalSharedTransitionApi::class)
-@Composable
-internal fun SharedTransitionScope.InteractiveShortcutInfoGridItem(
-    modifier: Modifier = Modifier,
-    data: GridItemData.ShortcutInfo,
-    drag: Drag,
-    gridItem: GridItem,
-    gridItemSettings: GridItemSettings,
-    hasShortcutHostPermission: Boolean,
-    isScrollInProgress: Boolean,
-    isSelected: Boolean,
-    isShowWhiteBox: Boolean,
-    isVisibleFolder: Boolean,
-    isVisibleOverlay: Boolean,
-    newGridItemSource: GridItemSource,
-    sharedElementKey: SharedElementKey,
-    textColor: Color,
-    onDismissGridItemPopup: () -> Unit,
-    onOpenAppDrawer: () -> Unit,
-    onShowGridItemPopup: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onTapShortcutInfo: (
-        serialNumber: Long,
-        packageName: String,
-        shortcutId: String,
-    ) -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
-) {
-    val launcherApps = LocalLauncherApps.current
-
-    val context = LocalContext.current
-
-    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
-
-    var intSize by remember { mutableStateOf(IntSize.Zero) }
-
-    val graphicsLayer = rememberGraphicsLayer()
-
-    val scope = rememberCoroutineScope()
-
-    val horizontalAlignment =
-        getHorizontalAlignment(horizontalAlignment = gridItemSettings.horizontalAlignment)
-
-    val verticalArrangement =
-        getVerticalArrangement(verticalArrangement = gridItemSettings.verticalArrangement)
-
-    val maxLines = if (gridItemSettings.singleLineLabel) 1 else Int.MAX_VALUE
-
-    val customIcon = data.customIcon ?: data.icon
-
-    val customShortLabel = data.customShortLabel ?: data.shortLabel
-
-    val hasInteraction = isSelected && isVisibleOverlay
-
-    val isVisibleWhiteBox = isShowWhiteBox && isSelected && drag == Drag.Dragging
-
-    val alpha = if (hasInteraction) 0f else 1f
-
-    LaunchedEffect(key1 = drag) {
-        handleDrag(
-            drag = drag,
-            isSelected = isSelected,
-            isVisibleOverlay = isVisibleOverlay,
-            onUpdateIsDragging = onUpdateIsDragging,
-            onDismissGridItemPopup = onDismissGridItemPopup,
-        )
-    }
-
-    Column(
-        modifier = modifier
-            .pointerInput(key1 = drag) {
-                detectTapGestures(
-                    onDoubleTap = if (!isVisibleOverlay) {
-                        {
-                            onDoubleTap(
-                                context = context,
-                                doubleTap = gridItem.doubleTap,
-                                launcherApps = launcherApps,
-                                scope = scope,
-                                onOpenAppDrawer = onOpenAppDrawer,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onLongPress = if (!isVisibleOverlay) {
-                        {
-                            onLongPress(
-                                scope = scope,
-                                graphicsLayer = graphicsLayer,
-                                intOffset = intOffset,
-                                intSize = intSize,
-                                gridItemSource = newGridItemSource,
-                                sharedElementKey = sharedElementKey,
-                                gridItem = gridItem,
-                                onUpdateGridItemSource = onUpdateGridItemSource,
-                                onUpdateImageBitmap = onUpdateImageBitmap,
-                                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                onShowGridItemPopup = onShowGridItemPopup,
-                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onTap = if (!isVisibleOverlay) {
-                        {
-                            if (hasShortcutHostPermission && data.isEnabled) {
-                                scope.launch {
-                                    onTapShortcutInfo(
-                                        data.serialNumber,
-                                        data.packageName,
-                                        data.shortcutId,
-                                    )
-                                }
-                            }
-                        }
-                    } else {
-                        null
-                    },
-                )
-            }
-            .swipeGestures(
-                swipeDown = gridItem.swipeDown,
-                swipeUp = gridItem.swipeUp,
-                onOpenAppDrawer = onOpenAppDrawer,
-            )
-            .fillMaxSize()
-            .padding(gridItemSettings.padding.dp)
-            .background(
-                color = Color(gridItemSettings.customBackgroundColor),
-                shape = RoundedCornerShape(size = gridItemSettings.cornerRadius.dp),
-            )
-            .whiteBox(
-                textColor = textColor,
-
-                visible = isVisibleWhiteBox && !isVisibleFolder,
-            ),
-        horizontalAlignment = horizontalAlignment,
-        verticalArrangement = verticalArrangement,
-    ) {
-        Box(
-            modifier = Modifier
-                .size(gridItemSettings.iconSize.dp)
-                .alpha(alpha),
-        ) {
-            AsyncImage(
-                model = Builder(LocalContext.current).data(customIcon)
-                    .addLastModifiedToFileCacheKey(true)
-                    .size(Size.ORIGINAL)
-                    .build(),
-                modifier = Modifier
-                    .matchParentSize()
-                    .drawWithContent {
-                        graphicsLayer.record {
-                            this@drawWithContent.drawContent()
-                        }
-
-                        drawLayer(graphicsLayer)
-                    }
-                    .onGloballyPositioned { layoutCoordinates ->
-                        intOffset = layoutCoordinates.positionInRoot().round()
-
-                        intSize = layoutCoordinates.size
-                    }
-                    .run {
-                        if (!hasInteraction) {
-                            sharedElementWithCallerManagedVisibility(
-                                rememberSharedContentState(
-                                    key = sharedElementKey,
-                                ),
-                                visible = !isScrollInProgress,
-                            )
-                        } else {
-                            this
-                        }
-                    },
-                contentDescription = null,
-            )
-
-            AsyncImage(
-                model = Builder(LocalContext.current).data(data.eblanApplicationInfoIcon)
-                    .size(Size.ORIGINAL)
-                    .build(),
-                modifier = Modifier
-                    .size((gridItemSettings.iconSize * 0.25).dp)
-                    .align(Alignment.BottomEnd),
-                contentDescription = null,
-            )
-        }
-
-        if (gridItemSettings.showLabel) {
-            Text(
-                modifier = Modifier.alpha(alpha),
-                text = customShortLabel,
-                color = textColor,
-                textAlign = TextAlign.Center,
-                maxLines = maxLines,
-                fontSize = gridItemSettings.textSize.sp,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalSharedTransitionApi::class)
-@Composable
-internal fun SharedTransitionScope.InteractiveFolderGridItem(
-    modifier: Modifier = Modifier,
-    data: GridItemData.Folder,
-    drag: Drag,
-    gridItem: GridItem,
-    gridItemSettings: GridItemSettings,
-    iconPackFilePaths: Map<String, String>,
-    isScrollInProgress: Boolean,
-    isSelected: Boolean,
-    isShowWhiteBox: Boolean,
-    isVisibleFolder: Boolean,
-    isVisibleOverlay: Boolean,
-    newGridItemSource: GridItemSource,
-    sharedElementKey: SharedElementKey,
-    textColor: Color,
-    moveGridItemResult: MoveGridItemResult?,
-    isDragging: Boolean,
-    lockMovement: Boolean,
-    onDismissGridItemPopup: () -> Unit,
-    onOpenAppDrawer: () -> Unit,
-    onShowGridItemPopup: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onTap: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
-    onShowFolderWhenDragging: (
-        conflictingGridItem: GridItem,
-        movingGridItem: GridItem,
-    ) -> Unit,
-    onUpdateFolderPopupBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-) {
-    val launcherApps = LocalLauncherApps.current
-
-    val context = LocalContext.current
-
-    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
-
-    var intSize by remember { mutableStateOf(IntSize.Zero) }
-
-    val graphicsLayer = rememberGraphicsLayer()
-
-    val scope = rememberCoroutineScope()
-
-    val horizontalAlignment =
-        getHorizontalAlignment(horizontalAlignment = gridItemSettings.horizontalAlignment)
-
-    val verticalArrangement =
-        getVerticalArrangement(verticalArrangement = gridItemSettings.verticalArrangement)
-
-    val maxLines = if (gridItemSettings.singleLineLabel) 1 else Int.MAX_VALUE
-
-    val hasInteraction = isSelected && isVisibleOverlay
-
-    val isVisibleWhiteBox = isShowWhiteBox && isSelected && drag == Drag.Dragging
-
-    val alpha = if (hasInteraction) 0f else 1f
-
-    LaunchedEffect(key1 = drag) {
-        handleDrag(
-            drag = drag,
-            isSelected = isSelected,
-            isVisibleOverlay = isVisibleOverlay,
-            onUpdateIsDragging = onUpdateIsDragging,
-            onDismissGridItemPopup = onDismissGridItemPopup,
-        )
-    }
-
-    LaunchedEffect(key1 = moveGridItemResult) {
-        handleConflictingGridItem(
-            drag = drag,
-            isDragging = isDragging,
-            isVisibleOverlay = isVisibleOverlay,
-            moveGridItemResult = moveGridItemResult,
-            lockMovement = lockMovement,
-            intOffset = intOffset,
-            intSize = intSize,
-            gridItem = gridItem,
-            onShowFolderWhenDragging = onShowFolderWhenDragging,
-            onUpdateFolderPopupBounds = onUpdateFolderPopupBounds,
-            onUpdateSharedElementKey = onUpdateSharedElementKey,
-        )
-    }
-
-    Column(
-        modifier = modifier
-            .pointerInput(key1 = drag) {
-                detectTapGestures(
-                    onDoubleTap = if (!isVisibleOverlay) {
-                        {
-                            onDoubleTap(
-                                context = context,
-                                doubleTap = gridItem.doubleTap,
-                                launcherApps = launcherApps,
-                                scope = scope,
-                                onOpenAppDrawer = onOpenAppDrawer,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onLongPress = if (!isVisibleOverlay) {
-                        {
-                            onLongPress(
-                                scope = scope,
-                                graphicsLayer = graphicsLayer,
-                                intOffset = intOffset,
-                                intSize = intSize,
-                                gridItemSource = newGridItemSource,
-                                sharedElementKey = sharedElementKey,
-                                gridItem = gridItem,
-                                onUpdateGridItemSource = onUpdateGridItemSource,
-                                onUpdateImageBitmap = onUpdateImageBitmap,
-                                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                onShowGridItemPopup = onShowGridItemPopup,
-                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onTap = if (!isVisibleOverlay) {
-                        {
-                            onTap(
-                                intOffset,
-                                intSize,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                )
-            }
-            .swipeGestures(
-                swipeDown = gridItem.swipeDown,
-                swipeUp = gridItem.swipeUp,
-                onOpenAppDrawer = onOpenAppDrawer,
-            )
-            .fillMaxSize()
-            .padding(gridItemSettings.padding.dp)
-            .background(
-                color = Color(gridItemSettings.customBackgroundColor),
-                shape = RoundedCornerShape(size = gridItemSettings.cornerRadius.dp),
-            )
-            .whiteBox(
-                textColor = textColor,
-                visible = isVisibleWhiteBox && !isVisibleFolder,
-            ),
-        horizontalAlignment = horizontalAlignment,
-        verticalArrangement = verticalArrangement,
-    ) {
-        val commonModifier = Modifier
-            .size(gridItemSettings.iconSize.dp)
-            .alpha(alpha)
-            .drawWithContent {
-                graphicsLayer.record {
-                    this@drawWithContent.drawContent()
-                }
-
-                drawLayer(graphicsLayer)
-            }
-            .onGloballyPositioned { layoutCoordinates ->
-                intOffset = layoutCoordinates.positionInRoot().round()
-
-                intSize = layoutCoordinates.size
-            }
-            .run {
-                if (!hasInteraction) {
-                    sharedElementWithCallerManagedVisibility(
-                        rememberSharedContentState(
-                            key = sharedElementKey,
-                        ),
-                        visible = !isScrollInProgress,
-                    )
-                } else {
-                    this
-                }
-            }
-
-        if (data.icon != null) {
-            AsyncImage(
-                model = data.icon,
-                contentDescription = null,
-                modifier = commonModifier,
-            )
-        } else {
-            Box(
-                modifier = commonModifier.background(
-                    color = MaterialTheme.colorScheme.surface.copy(alpha = 0.5f),
-                    shape = RoundedCornerShape(5.dp),
-                ),
-            ) {
-                FolderGridLayout(
-                    modifier = Modifier.matchParentSize(),
-                    columns = FOLDER_COLUMNS,
-                    gridItems = data.previewGridItemsByPage.take(FOLDER_COLUMNS * FOLDER_ROWS),
-                    rows = FOLDER_ROWS,
-                    content = { gridItem ->
-                        PreviewFolderGridItem(
-                            alpha = alpha,
-                            gridItem = gridItem,
-                            gridItemSettings = gridItemSettings,
-                            iconPackFilePaths = iconPackFilePaths,
-                            isScrollInProgress = isScrollInProgress,
-                            isVisibleOverlay = isVisibleOverlay,
-                            parent = sharedElementKey.parent,
-                            moveGridItemResult = moveGridItemResult,
-                        )
-                    },
-                )
-            }
-        }
-
-        if (gridItemSettings.showLabel) {
-            Text(
-                modifier = Modifier.alpha(alpha),
-                text = data.label,
-                color = textColor,
-                textAlign = TextAlign.Center,
-                maxLines = maxLines,
-                fontSize = gridItemSettings.textSize.sp,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalSharedTransitionApi::class)
-@Composable
-internal fun SharedTransitionScope.InteractiveShortcutConfigGridItem(
-    modifier: Modifier = Modifier,
-    data: GridItemData.ShortcutConfig,
-    drag: Drag,
-    gridItem: GridItem,
-    gridItemSettings: GridItemSettings,
-    isScrollInProgress: Boolean,
-    isSelected: Boolean,
-    isShowWhiteBox: Boolean,
-    isVisibleFolder: Boolean,
-    isVisibleOverlay: Boolean,
-    newGridItemSource: GridItemSource,
-    sharedElementKey: SharedElementKey,
-    textColor: Color,
-    onDismissGridItemPopup: () -> Unit,
-    onOpenAppDrawer: () -> Unit,
-    onShowGridItemPopup: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onTapShortcutConfig: (String) -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
-) {
-    val launcherApps = LocalLauncherApps.current
-
-    val context = LocalContext.current
-
-    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
-
-    var intSize by remember { mutableStateOf(IntSize.Zero) }
-
-    val graphicsLayer = rememberGraphicsLayer()
-
-    val scope = rememberCoroutineScope()
-
-    val horizontalAlignment =
-        getHorizontalAlignment(horizontalAlignment = gridItemSettings.horizontalAlignment)
-
-    val verticalArrangement =
-        getVerticalArrangement(verticalArrangement = gridItemSettings.verticalArrangement)
-
-    val maxLines = if (gridItemSettings.singleLineLabel) 1 else Int.MAX_VALUE
-
-    val icon = when {
-        data.customIcon != null -> {
-            data.customIcon
-        }
-
-        data.shortcutIntentIcon != null -> {
-            data.shortcutIntentIcon
-        }
-
-        data.activityIcon != null -> {
-            data.activityIcon
-        }
-
-        else -> {
-            data.applicationIcon
-        }
-    }
-
-    val label = when {
-        data.customLabel != null -> {
-            data.customLabel
-        }
-
-        data.shortcutIntentName != null -> {
-            data.shortcutIntentName
-        }
-
-        data.activityLabel != null -> {
-            data.activityLabel
-        }
-
-        else -> {
-            data.applicationLabel
-        }
-    }
-
-    val hasInteraction = isSelected && isVisibleOverlay
-
-    val isVisibleWhiteBox = isShowWhiteBox && isSelected && drag == Drag.Dragging
-
-    val alpha = if (hasInteraction) 0f else 1f
-
-    LaunchedEffect(key1 = drag) {
-        handleDrag(
-            drag = drag,
-            isSelected = isSelected,
-            isVisibleOverlay = isVisibleOverlay,
-            onUpdateIsDragging = onUpdateIsDragging,
-            onDismissGridItemPopup = onDismissGridItemPopup,
-        )
-    }
-
-    Column(
-        modifier = modifier
-            .pointerInput(key1 = drag) {
-                detectTapGestures(
-                    onDoubleTap = if (!isVisibleOverlay) {
-                        {
-                            onDoubleTap(
-                                context = context,
-                                doubleTap = gridItem.doubleTap,
-                                launcherApps = launcherApps,
-                                scope = scope,
-                                onOpenAppDrawer = onOpenAppDrawer,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onLongPress = if (!isVisibleOverlay) {
-                        {
-                            onLongPress(
-                                scope = scope,
-                                graphicsLayer = graphicsLayer,
-                                intOffset = intOffset,
-                                intSize = intSize,
-                                gridItemSource = newGridItemSource,
-                                sharedElementKey = sharedElementKey,
-                                gridItem = gridItem,
-                                onUpdateGridItemSource = onUpdateGridItemSource,
-                                onUpdateImageBitmap = onUpdateImageBitmap,
-                                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                onShowGridItemPopup = onShowGridItemPopup,
-                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
-                            )
-                        }
-                    } else {
-                        null
-                    },
-                    onTap = if (!isVisibleOverlay) {
-                        {
-                            data.shortcutIntentUri?.let(onTapShortcutConfig)
-                        }
-                    } else {
-                        null
-                    },
-                )
-            }
-            .swipeGestures(
-                swipeDown = gridItem.swipeDown,
-                swipeUp = gridItem.swipeUp,
-                onOpenAppDrawer = onOpenAppDrawer,
-            )
-            .fillMaxSize()
-            .padding(gridItemSettings.padding.dp)
-            .background(
-                color = Color(gridItemSettings.customBackgroundColor),
-                shape = RoundedCornerShape(size = gridItemSettings.cornerRadius.dp),
-            )
-            .whiteBox(
-                textColor = textColor,
-                visible = isVisibleWhiteBox && !isVisibleFolder,
-            ),
-        horizontalAlignment = horizontalAlignment,
-        verticalArrangement = verticalArrangement,
-    ) {
-        AsyncImage(
-            model = Builder(LocalContext.current).data(icon)
-                .addLastModifiedToFileCacheKey(true)
-                .size(Size.ORIGINAL)
-                .build(),
-            contentDescription = null,
-            modifier = Modifier
-                .size(gridItemSettings.iconSize.dp)
-                .alpha(alpha)
-                .drawWithContent {
-                    graphicsLayer.record {
-                        this@drawWithContent.drawContent()
-                    }
-
-                    drawLayer(graphicsLayer)
-                }
-                .onGloballyPositioned { layoutCoordinates ->
-                    intOffset = layoutCoordinates.positionInRoot().round()
-
-                    intSize = layoutCoordinates.size
-                }
-                .run {
-                    if (!hasInteraction) {
-                        sharedElementWithCallerManagedVisibility(
-                            rememberSharedContentState(
-                                key = sharedElementKey,
-                            ),
-                            visible = !isScrollInProgress,
-                        )
-                    } else {
-                        this
-                    }
-                },
-        )
-
-        if (gridItemSettings.showLabel) {
-            Text(
-                modifier = Modifier.alpha(alpha),
-                text = label.toString(),
-                color = textColor,
-                textAlign = TextAlign.Center,
-                maxLines = maxLines,
-                fontSize = gridItemSettings.textSize.sp,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
-
-@Composable
-private fun SharedTransitionScope.PreviewFolderGridItem(
-    modifier: Modifier = Modifier,
-    alpha: Float,
-    gridItem: GridItem,
-    gridItemSettings: GridItemSettings,
-    iconPackFilePaths: Map<String, String>,
-    isScrollInProgress: Boolean,
-    isVisibleOverlay: Boolean,
-    parent: SharedElementKey.Parent,
-    moveGridItemResult: MoveGridItemResult?,
-) {
-    val context = LocalContext.current
-
-    key(gridItem.id) {
-        val isSelected =
-            moveGridItemResult != null && moveGridItemResult.movingGridItem.id == gridItem.id
-
-        val hasInteraction = isSelected && isVisibleOverlay
-
-        val commonModifier = modifier
-            .alpha(alpha)
-            .size((gridItemSettings.iconSize * 0.30).dp)
-            .run {
-                if (!hasInteraction) {
-                    sharedElementWithCallerManagedVisibility(
-                        rememberSharedContentState(
-                            key = SharedElementKey(
-                                id = gridItem.id,
-                                parent = parent,
-                            ),
-                        ),
-                        visible = !isScrollInProgress,
-                    )
-                } else {
-                    this
-                }
-            }
-
-        when (val data = gridItem.data) {
-            is GridItemData.ApplicationInfo -> {
-                val icon =
-                    iconPackFilePaths[data.componentName]
-                        ?: data.icon
-
-                AsyncImage(
-                    model = Builder(context)
-                        .data(data.customIcon ?: icon)
-                        .addLastModifiedToFileCacheKey(true)
-                        .size(Size.ORIGINAL)
-                        .build(),
-                    contentDescription = null,
-                    modifier = commonModifier,
-                )
-            }
-
-            is GridItemData.ShortcutConfig -> {
-                val icon = when {
-                    data.customIcon != null -> {
-                        data.customIcon
-                    }
-
-                    data.shortcutIntentIcon != null -> {
-                        data.shortcutIntentIcon
-                    }
-
-                    data.activityIcon != null -> {
-                        data.activityIcon
-                    }
-
-                    else -> {
-                        data.applicationIcon
-                    }
-                }
-
-                AsyncImage(
-                    model = Builder(context)
-                        .data(icon)
-                        .addLastModifiedToFileCacheKey(true)
-                        .size(Size.ORIGINAL)
-                        .build(),
-                    contentDescription = null,
-                    modifier = commonModifier,
-                )
-            }
-
-            is GridItemData.ShortcutInfo -> {
-                AsyncImage(
-                    model = Builder(context)
-                        .data(data.customIcon ?: data.icon)
-                        .addLastModifiedToFileCacheKey(true)
-                        .size(Size.ORIGINAL)
-                        .build(),
-                    contentDescription = null,
-                    modifier = commonModifier,
-                )
-            }
-
-            else -> Unit
-        }
-    }
-}
