+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/folder/InteractiveFolderGridItem.kt
@@ -19,23 +19,65 @@ package com.eblan.launcher.feature.home.screen.folder
 
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionScope
+import androidx.compose.foundation.background
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.layer.drawLayer
+import androidx.compose.ui.graphics.rememberGraphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.layout.positionInRoot
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.lerp
+import androidx.compose.ui.unit.round
+import androidx.compose.ui.unit.sp
+import coil3.compose.AsyncImage
+import coil3.request.ImageRequest.Builder
+import coil3.request.addLastModifiedToFileCacheKey
+import coil3.size.Size
 import com.eblan.launcher.domain.model.GridItem
 import com.eblan.launcher.domain.model.GridItemData
 import com.eblan.launcher.domain.model.GridItemSettings
 import com.eblan.launcher.domain.model.MoveGridItemResult
-import com.eblan.launcher.feature.home.component.InteractiveApplicationInfoGridItem
-import com.eblan.launcher.feature.home.component.InteractiveFolderGridItem
-import com.eblan.launcher.feature.home.component.InteractiveShortcutConfigGridItem
-import com.eblan.launcher.feature.home.component.InteractiveShortcutInfoGridItem
+import com.eblan.launcher.feature.home.component.swipeGestures
 import com.eblan.launcher.feature.home.model.Drag
 import com.eblan.launcher.feature.home.model.GridItemSource
 import com.eblan.launcher.feature.home.model.SharedElementKey
+import com.eblan.launcher.feature.home.util.getHorizontalAlignment
+import com.eblan.launcher.feature.home.util.getVerticalArrangement
+import com.eblan.launcher.feature.home.util.handleDrag
+import com.eblan.launcher.feature.home.util.onDoubleTap
+import com.eblan.launcher.feature.home.util.onLongPress
+import com.eblan.launcher.ui.local.LocalLauncherApps
+import com.eblan.launcher.ui.local.LocalSettings
+import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalSharedTransitionApi::class)
 @Composable
@@ -52,8 +94,7 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
     newGridItemSource: GridItemSource,
     sharedElementKey: SharedElementKey,
     moveGridItemResult: MoveGridItemResult?,
-    isDragging: Boolean,
-    lockMovement: Boolean,
+    progress: Float,
     onOpenAppDrawer: () -> Unit,
     onTapApplicationInfo: (
         serialNumber: Long,
@@ -80,14 +121,6 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
     onDismissGridItemPopup: () -> Unit,
     onUpdateIsVisibleOverlay: (Boolean) -> Unit,
     onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
-    onShowFolderWhenDragging: (
-        conflictingGridItem: GridItem,
-        movingGridItem: GridItem,
-    ) -> Unit,
-    onUpdateFolderPopupBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
 ) {
     val isSelected =
         moveGridItemResult != null && moveGridItemResult.movingGridItem.id == gridItem.id
@@ -98,9 +131,11 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
         gridItemSettings
     }
 
+    val padding = lerp(1.dp, gridItemSettings.padding.dp, progress)
+
     when (val data = gridItem.data) {
         is GridItemData.ApplicationInfo -> {
-            InteractiveApplicationInfoGridItem(
+            InteractiveFolderApplicationInfoGridItem(
                 modifier = modifier,
                 data = data,
                 drag = drag,
@@ -109,13 +144,11 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
                 iconPackFilePaths = iconPackFilePaths,
                 isScrollInProgress = isScrollInProgress,
                 isSelected = isSelected,
-                isShowWhiteBox = false,
-                isVisibleFolder = false,
                 isVisibleOverlay = isVisibleOverlay,
                 newGridItemSource = newGridItemSource,
                 sharedElementKey = sharedElementKey,
                 statusBarNotifications = statusBarNotifications,
-                textColor = Color.Unspecified,
+                padding = padding,
                 onDismissGridItemPopup = onDismissGridItemPopup,
                 onOpenAppDrawer = onOpenAppDrawer,
                 onShowGridItemPopup = onShowGridItemPopup,
@@ -130,10 +163,8 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
             )
         }
 
-        is GridItemData.Widget -> error("Unsupported Folder Grid Item")
-
         is GridItemData.ShortcutInfo -> {
-            InteractiveShortcutInfoGridItem(
+            InteractiveFolderShortcutInfoGridItem(
                 modifier = modifier,
                 data = data,
                 drag = drag,
@@ -142,12 +173,10 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
                 hasShortcutHostPermission = hasShortcutHostPermission,
                 isScrollInProgress = isScrollInProgress,
                 isSelected = isSelected,
-                isShowWhiteBox = false,
-                isVisibleFolder = false,
                 isVisibleOverlay = isVisibleOverlay,
                 newGridItemSource = newGridItemSource,
                 sharedElementKey = sharedElementKey,
-                textColor = Color.Unspecified,
+                padding = padding,
                 onDismissGridItemPopup = onDismissGridItemPopup,
                 onOpenAppDrawer = onOpenAppDrawer,
                 onShowGridItemPopup = onShowGridItemPopup,
@@ -162,44 +191,8 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
             )
         }
 
-        is GridItemData.Folder -> {
-            InteractiveFolderGridItem(
-                modifier = modifier,
-                data = data,
-                drag = drag,
-                gridItem = gridItem,
-                gridItemSettings = currentGridItemSettings,
-                iconPackFilePaths = iconPackFilePaths,
-                isScrollInProgress = isScrollInProgress,
-                isSelected = isSelected,
-                isShowWhiteBox = false,
-                isVisibleFolder = false,
-                isVisibleOverlay = isVisibleOverlay,
-                newGridItemSource = newGridItemSource,
-                sharedElementKey = sharedElementKey,
-                textColor = Color.Unspecified,
-                moveGridItemResult = moveGridItemResult,
-                isDragging = isDragging,
-                lockMovement = lockMovement,
-                onDismissGridItemPopup = onDismissGridItemPopup,
-                onOpenAppDrawer = onOpenAppDrawer,
-                onShowGridItemPopup = onShowGridItemPopup,
-                onTap = { _, _ -> },
-                onUpdateGridItemSource = onUpdateGridItemSource,
-                onUpdateImageBitmap = onUpdateImageBitmap,
-                onUpdateIsDragging = onUpdateIsDragging,
-                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
-                onShowFolderWhenDragging = onShowFolderWhenDragging,
-                onUpdateFolderPopupBounds = onUpdateFolderPopupBounds,
-
-            )
-        }
-
         is GridItemData.ShortcutConfig -> {
-            InteractiveShortcutConfigGridItem(
+            InteractiveFolderShortcutConfigGridItem(
                 modifier = modifier,
                 data = data,
                 drag = drag,
@@ -207,12 +200,10 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
                 gridItemSettings = currentGridItemSettings,
                 isScrollInProgress = isScrollInProgress,
                 isSelected = isSelected,
-                isShowWhiteBox = false,
-                isVisibleFolder = false,
                 isVisibleOverlay = isVisibleOverlay,
                 newGridItemSource = newGridItemSource,
                 sharedElementKey = sharedElementKey,
-                textColor = Color.Unspecified,
+                padding = padding,
                 onDismissGridItemPopup = onDismissGridItemPopup,
                 onOpenAppDrawer = onOpenAppDrawer,
                 onShowGridItemPopup = onShowGridItemPopup,
@@ -224,6 +215,643 @@ internal fun SharedTransitionScope.InteractiveFolderGridItemContent(
                 onUpdateOverlayBounds = onUpdateOverlayBounds,
                 onUpdateSharedElementKey = onUpdateSharedElementKey,
                 onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
+            )
+        }
+
+        is GridItemData.Widget,
+        is GridItemData.Folder,
+        -> error("Unsupported Folder Grid Item")
+    }
+}
+
+@OptIn(ExperimentalSharedTransitionApi::class)
+@Composable
+private fun SharedTransitionScope.InteractiveFolderApplicationInfoGridItem(
+    modifier: Modifier = Modifier,
+    data: GridItemData.ApplicationInfo,
+    drag: Drag,
+    gridItem: GridItem,
+    gridItemSettings: GridItemSettings,
+    iconPackFilePaths: Map<String, String>,
+    isScrollInProgress: Boolean,
+    isSelected: Boolean,
+    isVisibleOverlay: Boolean,
+    newGridItemSource: GridItemSource,
+    sharedElementKey: SharedElementKey,
+    statusBarNotifications: Map<String, Int>,
+    padding: Dp,
+    onDismissGridItemPopup: () -> Unit,
+    onOpenAppDrawer: () -> Unit,
+    onShowGridItemPopup: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onTapApplicationInfo: (
+        serialNumber: Long,
+        componentName: String,
+    ) -> Unit,
+    onUpdateGridItemSource: (GridItemSource) -> Unit,
+    onUpdateImageBitmap: (ImageBitmap) -> Unit,
+    onUpdateIsDragging: (Boolean) -> Unit,
+    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
+    onUpdateOverlayBounds: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
+    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
+) {
+    val launcherApps = LocalLauncherApps.current
+
+    val context = LocalContext.current
+
+    val settings = LocalSettings.current
+
+    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
+
+    var intSize by remember { mutableStateOf(IntSize.Zero) }
+
+    val graphicsLayer = rememberGraphicsLayer()
+
+    val scope = rememberCoroutineScope()
+
+    val horizontalAlignment =
+        getHorizontalAlignment(horizontalAlignment = gridItemSettings.horizontalAlignment)
+
+    val verticalArrangement =
+        getVerticalArrangement(verticalArrangement = gridItemSettings.verticalArrangement)
+
+    val maxLines = if (gridItemSettings.singleLineLabel) 1 else Int.MAX_VALUE
+
+    val icon = iconPackFilePaths[data.componentName] ?: data.icon
+
+    val hasNotifications =
+        statusBarNotifications[data.packageName] != null && (
+            statusBarNotifications[data.packageName]
+                ?: 0
+            ) > 0
+
+    val hasInteraction = isSelected && isVisibleOverlay
+
+    val alpha = if (hasInteraction) 0f else 1f
+
+    LaunchedEffect(key1 = drag) {
+        handleDrag(
+            drag = drag,
+            isSelected = isSelected,
+            isVisibleOverlay = isVisibleOverlay,
+            onUpdateIsDragging = onUpdateIsDragging,
+            onDismissGridItemPopup = onDismissGridItemPopup,
+        )
+    }
+
+    Column(
+        modifier = modifier
+            .pointerInput(key1 = drag) {
+                detectTapGestures(
+                    onDoubleTap = if (!isVisibleOverlay) {
+                        {
+                            onDoubleTap(
+                                context = context,
+                                doubleTap = gridItem.doubleTap,
+                                launcherApps = launcherApps,
+                                scope = scope,
+                                onOpenAppDrawer = onOpenAppDrawer,
+                            )
+                        }
+                    } else {
+                        null
+                    },
+                    onLongPress = if (!isVisibleOverlay) {
+                        {
+                            onLongPress(
+                                scope = scope,
+                                graphicsLayer = graphicsLayer,
+                                intOffset = intOffset,
+                                intSize = intSize,
+                                gridItemSource = newGridItemSource,
+                                sharedElementKey = sharedElementKey,
+                                gridItem = gridItem,
+                                onUpdateGridItemSource = onUpdateGridItemSource,
+                                onUpdateImageBitmap = onUpdateImageBitmap,
+                                onUpdateOverlayBounds = onUpdateOverlayBounds,
+                                onUpdateSharedElementKey = onUpdateSharedElementKey,
+                                onShowGridItemPopup = onShowGridItemPopup,
+                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
+                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
+                            )
+                        }
+                    } else {
+                        null
+                    },
+                    onTap = if (!isVisibleOverlay) {
+                        {
+                            scope.launch {
+                                onTapApplicationInfo(
+                                    data.serialNumber,
+                                    data.componentName,
+                                )
+                            }
+                        }
+                    } else {
+                        null
+                    },
+                )
+            }
+            .swipeGestures(
+                swipeDown = gridItem.swipeDown,
+                swipeUp = gridItem.swipeUp,
+                onOpenAppDrawer = onOpenAppDrawer,
+            )
+            .fillMaxSize()
+            .padding(padding)
+            .background(
+                color = Color(gridItemSettings.customBackgroundColor),
+                shape = RoundedCornerShape(size = gridItemSettings.cornerRadius.dp),
+            ),
+        horizontalAlignment = horizontalAlignment,
+        verticalArrangement = verticalArrangement,
+    ) {
+        Box(
+            modifier = Modifier
+                .size(gridItemSettings.iconSize.dp)
+                .alpha(alpha),
+        ) {
+            AsyncImage(
+                model = Builder(LocalContext.current).data(data.customIcon ?: icon)
+                    .addLastModifiedToFileCacheKey(true)
+                    .size(Size.ORIGINAL)
+                    .build(),
+                contentDescription = null,
+                modifier = Modifier
+                    .matchParentSize()
+                    .drawWithContent {
+                        graphicsLayer.record {
+                            this@drawWithContent.drawContent()
+                        }
+
+                        drawLayer(graphicsLayer)
+                    }
+                    .onGloballyPositioned { layoutCoordinates ->
+                        intOffset = layoutCoordinates.positionInRoot().round()
+
+                        intSize = layoutCoordinates.size
+                    }
+                    .run {
+                        if (!hasInteraction) {
+                            sharedElementWithCallerManagedVisibility(
+                                rememberSharedContentState(
+                                    key = sharedElementKey,
+                                ),
+                                visible = !isScrollInProgress,
+                            )
+                        } else {
+                            this
+                        }
+                    },
+            )
+
+            if (settings.isNotificationAccessGranted() && hasNotifications) {
+                Box(
+                    modifier = Modifier
+                        .size((gridItemSettings.iconSize * 0.3).dp)
+                        .align(Alignment.TopEnd)
+                        .background(
+                            color = MaterialTheme.colorScheme.primary,
+                            shape = CircleShape,
+                        ),
+                )
+            }
+        }
+
+        if (gridItemSettings.showLabel) {
+            Text(
+                modifier = Modifier.alpha(alpha),
+                text = data.customLabel ?: data.label,
+                textAlign = TextAlign.Center,
+                maxLines = maxLines,
+                fontSize = gridItemSettings.textSize.sp,
+                overflow = TextOverflow.Ellipsis,
+            )
+        }
+    }
+}
+
+@OptIn(ExperimentalSharedTransitionApi::class)
+@Composable
+private fun SharedTransitionScope.InteractiveFolderShortcutInfoGridItem(
+    modifier: Modifier = Modifier,
+    data: GridItemData.ShortcutInfo,
+    drag: Drag,
+    gridItem: GridItem,
+    gridItemSettings: GridItemSettings,
+    hasShortcutHostPermission: Boolean,
+    isScrollInProgress: Boolean,
+    isSelected: Boolean,
+    isVisibleOverlay: Boolean,
+    newGridItemSource: GridItemSource,
+    sharedElementKey: SharedElementKey,
+    padding: Dp,
+    onDismissGridItemPopup: () -> Unit,
+    onOpenAppDrawer: () -> Unit,
+    onShowGridItemPopup: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onTapShortcutInfo: (
+        serialNumber: Long,
+        packageName: String,
+        shortcutId: String,
+    ) -> Unit,
+    onUpdateGridItemSource: (GridItemSource) -> Unit,
+    onUpdateImageBitmap: (ImageBitmap) -> Unit,
+    onUpdateIsDragging: (Boolean) -> Unit,
+    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
+    onUpdateOverlayBounds: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
+    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
+) {
+    val launcherApps = LocalLauncherApps.current
+
+    val context = LocalContext.current
+
+    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
+
+    var intSize by remember { mutableStateOf(IntSize.Zero) }
+
+    val graphicsLayer = rememberGraphicsLayer()
+
+    val scope = rememberCoroutineScope()
+
+    val horizontalAlignment =
+        getHorizontalAlignment(horizontalAlignment = gridItemSettings.horizontalAlignment)
+
+    val verticalArrangement =
+        getVerticalArrangement(verticalArrangement = gridItemSettings.verticalArrangement)
+
+    val maxLines = if (gridItemSettings.singleLineLabel) 1 else Int.MAX_VALUE
+
+    val customIcon = data.customIcon ?: data.icon
+
+    val customShortLabel = data.customShortLabel ?: data.shortLabel
+
+    val hasInteraction = isSelected && isVisibleOverlay
+
+    val alpha = if (hasInteraction) 0f else 1f
+
+    LaunchedEffect(key1 = drag) {
+        handleDrag(
+            drag = drag,
+            isSelected = isSelected,
+            isVisibleOverlay = isVisibleOverlay,
+            onUpdateIsDragging = onUpdateIsDragging,
+            onDismissGridItemPopup = onDismissGridItemPopup,
+        )
+    }
+
+    Column(
+        modifier = modifier
+            .pointerInput(key1 = drag) {
+                detectTapGestures(
+                    onDoubleTap = if (!isVisibleOverlay) {
+                        {
+                            onDoubleTap(
+                                context = context,
+                                doubleTap = gridItem.doubleTap,
+                                launcherApps = launcherApps,
+                                scope = scope,
+                                onOpenAppDrawer = onOpenAppDrawer,
+                            )
+                        }
+                    } else {
+                        null
+                    },
+                    onLongPress = if (!isVisibleOverlay) {
+                        {
+                            onLongPress(
+                                scope = scope,
+                                graphicsLayer = graphicsLayer,
+                                intOffset = intOffset,
+                                intSize = intSize,
+                                gridItemSource = newGridItemSource,
+                                sharedElementKey = sharedElementKey,
+                                gridItem = gridItem,
+                                onUpdateGridItemSource = onUpdateGridItemSource,
+                                onUpdateImageBitmap = onUpdateImageBitmap,
+                                onUpdateOverlayBounds = onUpdateOverlayBounds,
+                                onUpdateSharedElementKey = onUpdateSharedElementKey,
+                                onShowGridItemPopup = onShowGridItemPopup,
+                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
+                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
+                            )
+                        }
+                    } else {
+                        null
+                    },
+                    onTap = if (!isVisibleOverlay) {
+                        {
+                            if (hasShortcutHostPermission && data.isEnabled) {
+                                scope.launch {
+                                    onTapShortcutInfo(
+                                        data.serialNumber,
+                                        data.packageName,
+                                        data.shortcutId,
+                                    )
+                                }
+                            }
+                        }
+                    } else {
+                        null
+                    },
+                )
+            }
+            .swipeGestures(
+                swipeDown = gridItem.swipeDown,
+                swipeUp = gridItem.swipeUp,
+                onOpenAppDrawer = onOpenAppDrawer,
+            )
+            .fillMaxSize()
+            .padding(padding)
+            .background(
+                color = Color(gridItemSettings.customBackgroundColor),
+                shape = RoundedCornerShape(size = gridItemSettings.cornerRadius.dp),
+            ),
+        horizontalAlignment = horizontalAlignment,
+        verticalArrangement = verticalArrangement,
+    ) {
+        Box(
+            modifier = Modifier
+                .size(gridItemSettings.iconSize.dp)
+                .alpha(alpha),
+        ) {
+            AsyncImage(
+                model = Builder(LocalContext.current).data(customIcon)
+                    .addLastModifiedToFileCacheKey(true)
+                    .size(Size.ORIGINAL)
+                    .build(),
+                modifier = Modifier
+                    .matchParentSize()
+                    .drawWithContent {
+                        graphicsLayer.record {
+                            this@drawWithContent.drawContent()
+                        }
+
+                        drawLayer(graphicsLayer)
+                    }
+                    .onGloballyPositioned { layoutCoordinates ->
+                        intOffset = layoutCoordinates.positionInRoot().round()
+
+                        intSize = layoutCoordinates.size
+                    }
+                    .run {
+                        if (!hasInteraction) {
+                            sharedElementWithCallerManagedVisibility(
+                                rememberSharedContentState(
+                                    key = sharedElementKey,
+                                ),
+                                visible = !isScrollInProgress,
+                            )
+                        } else {
+                            this
+                        }
+                    },
+                contentDescription = null,
+            )
+
+            AsyncImage(
+                model = Builder(LocalContext.current).data(data.eblanApplicationInfoIcon)
+                    .size(Size.ORIGINAL)
+                    .build(),
+                modifier = Modifier
+                    .size((gridItemSettings.iconSize * 0.25).dp)
+                    .align(Alignment.BottomEnd),
+                contentDescription = null,
+            )
+        }
+
+        if (gridItemSettings.showLabel) {
+            Text(
+                modifier = Modifier.alpha(alpha),
+                text = customShortLabel,
+                textAlign = TextAlign.Center,
+                maxLines = maxLines,
+                fontSize = gridItemSettings.textSize.sp,
+                overflow = TextOverflow.Ellipsis,
+            )
+        }
+    }
+}
+
+@OptIn(ExperimentalSharedTransitionApi::class)
+@Composable
+private fun SharedTransitionScope.InteractiveFolderShortcutConfigGridItem(
+    modifier: Modifier = Modifier,
+    data: GridItemData.ShortcutConfig,
+    drag: Drag,
+    gridItem: GridItem,
+    gridItemSettings: GridItemSettings,
+    isScrollInProgress: Boolean,
+    isSelected: Boolean,
+    isVisibleOverlay: Boolean,
+    newGridItemSource: GridItemSource,
+    sharedElementKey: SharedElementKey,
+    padding: Dp,
+    onDismissGridItemPopup: () -> Unit,
+    onOpenAppDrawer: () -> Unit,
+    onShowGridItemPopup: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onTapShortcutConfig: (String) -> Unit,
+    onUpdateGridItemSource: (GridItemSource) -> Unit,
+    onUpdateImageBitmap: (ImageBitmap) -> Unit,
+    onUpdateIsDragging: (Boolean) -> Unit,
+    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
+    onUpdateOverlayBounds: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
+    onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
+) {
+    val launcherApps = LocalLauncherApps.current
+
+    val context = LocalContext.current
+
+    var intOffset by remember { mutableStateOf(IntOffset.Zero) }
+
+    var intSize by remember { mutableStateOf(IntSize.Zero) }
+
+    val graphicsLayer = rememberGraphicsLayer()
+
+    val scope = rememberCoroutineScope()
+
+    val horizontalAlignment =
+        getHorizontalAlignment(horizontalAlignment = gridItemSettings.horizontalAlignment)
+
+    val verticalArrangement =
+        getVerticalArrangement(verticalArrangement = gridItemSettings.verticalArrangement)
+
+    val maxLines = if (gridItemSettings.singleLineLabel) 1 else Int.MAX_VALUE
+
+    val icon = when {
+        data.customIcon != null -> {
+            data.customIcon
+        }
+
+        data.shortcutIntentIcon != null -> {
+            data.shortcutIntentIcon
+        }
+
+        data.activityIcon != null -> {
+            data.activityIcon
+        }
+
+        else -> {
+            data.applicationIcon
+        }
+    }
+
+    val label = when {
+        data.customLabel != null -> {
+            data.customLabel
+        }
+
+        data.shortcutIntentName != null -> {
+            data.shortcutIntentName
+        }
+
+        data.activityLabel != null -> {
+            data.activityLabel
+        }
+
+        else -> {
+            data.applicationLabel
+        }
+    }
+
+    val hasInteraction = isSelected && isVisibleOverlay
+
+    val alpha = if (hasInteraction) 0f else 1f
+
+    LaunchedEffect(key1 = drag) {
+        handleDrag(
+            drag = drag,
+            isSelected = isSelected,
+            isVisibleOverlay = isVisibleOverlay,
+            onUpdateIsDragging = onUpdateIsDragging,
+            onDismissGridItemPopup = onDismissGridItemPopup,
+        )
+    }
+
+    Column(
+        modifier = modifier
+            .pointerInput(key1 = drag) {
+                detectTapGestures(
+                    onDoubleTap = if (!isVisibleOverlay) {
+                        {
+                            onDoubleTap(
+                                context = context,
+                                doubleTap = gridItem.doubleTap,
+                                launcherApps = launcherApps,
+                                scope = scope,
+                                onOpenAppDrawer = onOpenAppDrawer,
+                            )
+                        }
+                    } else {
+                        null
+                    },
+                    onLongPress = if (!isVisibleOverlay) {
+                        {
+                            onLongPress(
+                                scope = scope,
+                                graphicsLayer = graphicsLayer,
+                                intOffset = intOffset,
+                                intSize = intSize,
+                                gridItemSource = newGridItemSource,
+                                sharedElementKey = sharedElementKey,
+                                gridItem = gridItem,
+                                onUpdateGridItemSource = onUpdateGridItemSource,
+                                onUpdateImageBitmap = onUpdateImageBitmap,
+                                onUpdateOverlayBounds = onUpdateOverlayBounds,
+                                onUpdateSharedElementKey = onUpdateSharedElementKey,
+                                onShowGridItemPopup = onShowGridItemPopup,
+                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
+                                onUpdateMoveGridItemResult = onUpdateMoveGridItemResult,
+                            )
+                        }
+                    } else {
+                        null
+                    },
+                    onTap = if (!isVisibleOverlay) {
+                        {
+                            data.shortcutIntentUri?.let(onTapShortcutConfig)
+                        }
+                    } else {
+                        null
+                    },
+                )
+            }
+            .swipeGestures(
+                swipeDown = gridItem.swipeDown,
+                swipeUp = gridItem.swipeUp,
+                onOpenAppDrawer = onOpenAppDrawer,
+            )
+            .fillMaxSize()
+            .padding(padding)
+            .background(
+                color = Color(gridItemSettings.customBackgroundColor),
+                shape = RoundedCornerShape(size = gridItemSettings.cornerRadius.dp),
+            ),
+        horizontalAlignment = horizontalAlignment,
+        verticalArrangement = verticalArrangement,
+    ) {
+        AsyncImage(
+            model = Builder(LocalContext.current).data(icon)
+                .addLastModifiedToFileCacheKey(true)
+                .size(Size.ORIGINAL)
+                .build(),
+            contentDescription = null,
+            modifier = Modifier
+                .size(gridItemSettings.iconSize.dp)
+                .alpha(alpha)
+                .drawWithContent {
+                    graphicsLayer.record {
+                        this@drawWithContent.drawContent()
+                    }
+
+                    drawLayer(graphicsLayer)
+                }
+                .onGloballyPositioned { layoutCoordinates ->
+                    intOffset = layoutCoordinates.positionInRoot().round()
+
+                    intSize = layoutCoordinates.size
+                }
+                .run {
+                    if (!hasInteraction) {
+                        sharedElementWithCallerManagedVisibility(
+                            rememberSharedContentState(
+                                key = sharedElementKey,
+                            ),
+                            visible = !isScrollInProgress,
+                        )
+                    } else {
+                        this
+                    }
+                },
+        )
+
+        if (gridItemSettings.showLabel) {
+            Text(
+                modifier = Modifier.alpha(alpha),
+                text = label.toString(),
+                textAlign = TextAlign.Center,
+                maxLines = maxLines,
+                fontSize = gridItemSettings.textSize.sp,
+                overflow = TextOverflow.Ellipsis,
             )
         }
     }
